requirements.txt
pdf2image
Pillow
google-generativeai
vobject
pykakasi
python-dotenv
pytesseract
